Fix version of k8s provider
The given version of the k8s provider was not compatible with the
actual code in the module (running `terraform plan` revealed a lot of
error messages about version troubles).

We set the provider version to the newest version supporting the
`load_config_file` argument in the provider config.
Now, running `terraform plan` results in a successful plan with
no changes.

Internal ref [STRATUS-1091]

Co-authored-by: nomx3 <aet@ssb.no>
Co-authored-by: ssbdif <35797607+ssbdif@users.noreply.github.com>