Remove provider configuration
Provider configuration should be set in the root module.
Updated dependent module versions.

Keeping the "kubernetes" provider because of a dependency
with the gke sub-module.

The reason for changing this now, is because we need to
support Terraform 1.x for running locally. Only 1.x is
supported on Macs with Apple silicon.

Co-authored-by: ssbdif <35797607+ssbdif@users.noreply.github.com>