Expose variable to activate APIs in the project
This variable couldn't be used from the outside (for example by
`atlantis-prod`), since it was only exposed internally from the gcp
module. Now it's exposed externally as well and can be used by
`atlantis-prod` to specify which APIs should be enabled in its GCP
project so we can do that in code instead of manually in the GCP
console.